added a review database added a review class added list adapters: movies and reviews
changed id names to fit for listadapters
moshi

+++ app/src/main/java/edu/orangecoastcollege/cs220/dreyna3/freshlysqueezed/Movie.java
@@ -8,7 +8,7 @@ public class Movie {
 
     private String movieTitle;
     private String genre;
-    private int rating;
+    private float rating;
     private String description;
     private String imageName;
 
@@ -25,7 +25,7 @@ public class Movie {
         return genre;
     }
 
-    public int getRating() {
+    public float getRating() {
         return rating;
     }
